rag model/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
langchain-community==0.0.10
langchain-huggingface==0.0.1
chromadb==0.4.22
groq==0.4.2
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
